Split large image into the tiles (due to the max texture size limitation)

+++ interfaces/IVideoContext.hpp
@@ -11,8 +11,8 @@ public:
   virtual void setCamera(const Coord &) noexcept = 0;
   virtual void setScale(float) noexcept = 0;
   virtual void setAngle(double) noexcept = 0;
-  virtual int getWidth() const noexcept = 0;
-  virtual int getHeigt() const noexcept = 0;
+  virtual size_t getWidth() const noexcept = 0;
+  virtual size_t getHeigt() const noexcept = 0;
   virtual float getFps() const noexcept = 0;
   virtual void delay(size_t) const noexcept = 0;
   virtual size_t getMaxTextureSize() const noexcept = 0;
